+++ packages/firebase_dynamic_links/CHANGELOG.md
@@ -4,7 +4,7 @@
 the dynamic link which opened the app and links clicked during app execution (active and background).
 `retrieveDynamicLink` has been replaced with two different functions:
 - `getLaunchLink` a future to retrieve the link that opened the app
-- `configure` a callback to listen to links opened while the app is active or in background
+- `onLink` a callback to listen to links opened while the app is active or in background
 
 ## 0.4.0+5